requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
vtracer>=0.6.0
Pillow>=10.0.0
gunicorn>=21.0.0
